requirements.txt
pydantic
python-dotenv
wikipedia-api
groq
streamlit
pandas
openai